+++ veewee/scientificlinux-7.2-x86_64-netinst/ansible.sh
@@ -1,3 +1,0 @@
-
-# Install Ansible from EPEL: see epel.sh.
-yum -y install ansible